refactor: Add route for retrieving messages by ride ID
This commit adds a new route in the `socket/messages` module to handle GET requests for retrieving messages by ride ID. The route retrieves the ride ID from the request query parameters and uses it to fetch the corresponding ride from the database. If the ride is found, the route responds with a JSON object containing the success status and the messages associated with the ride. If the ride is not found, a 404 status is returned. If there is an error during the process, a 500 status is returned along with an error message.